+++ NEWS.md
@@ -1,8 +1,9 @@
 
-# rquery 0.2.0 2017/12/30
+# rquery 0.2.0 2017/12/31
 
  * Don't store DB reference.
  * Prepare for wrapr 1.1.0 features.
+ * Ad-hoc query modes.
   
 # rquery 0.1.0 2017/12/03